requirements.txt
openai==0.27.0
python-pptx==0.6.21
python-dotenv==1.0.0
azure-identity==1.7.0
